tests/wpt/tests/css/css-ui/parsing/outline-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: outline sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#outline">
<link rel="help" href="https://drafts.csswg.org/cssom/#serializing-css-values">
<meta name="assert" content="outline supports the full grammar '<outline-color> || <outline> || <outline>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('outline', '3px ridge blue', {
  'outline-color': 'blue',
  'outline-style': 'ridge',
  'outline-width': '3px'
});
</script>
</body>
</html>
